move test cert creation to testutils
